abstraction of common data for startDeviceMission, new configuration screen controller, refactor of all manage views, translate of all manage views,

+++ src/main/java/temperatus/controller/manage/ManageAuthorController.java
@@ -17,6 +17,7 @@ import temperatus.controller.AbstractController;
 import temperatus.controller.creation.NewAuthorController;
 import temperatus.lang.Lang;
 import temperatus.model.pojo.Author;
+import temperatus.model.pojo.Mission;
 import temperatus.model.service.AuthorService;
 import temperatus.util.Animation;
 import temperatus.util.Constants;
@@ -27,6 +28,8 @@ import java.util.Optional;
 import java.util.ResourceBundle;
 
 /**
+ * Allow the user to search, edit, delete and create authors
+ * <p>
  * Created by alberto on 15/2/16.
  */
 @Controller
@@ -38,52 +41,52 @@ public class ManageAuthorController implements Initializable, AbstractController
     @FXML private AnchorPane infoPane;
     @FXML private Button newElementButton;
 
+    @FXML private Button editAuthorButton;
+    @FXML private Button deleteAuthorButton;
+
+    @FXML private Label projectsLabel;
+    @FXML private Label nameInfo;
+    @FXML private Label projectsInfo;
+
     private TableColumn<Author, String> name = new TableColumn<>();
 
     private ObservableList<Author> authors;
 
     @Autowired AuthorService authorService;
 
-    static Logger logger = LoggerFactory.getLogger(ManageAuthorController.class.getName());
+    private static final String NEW_LINE = "\n";
+    private static Logger logger = LoggerFactory.getLogger(ManageAuthorController.class.getName());
 
     @Override
     public void initialize(URL location, ResourceBundle resources) {
         VistaNavigator.setController(this);
         translate();
 
-        authors = FXCollections.observableArrayList();
-        addAllAuthors();
-
-        name.setText("Name");
+        authors = FXCollections.observableArrayList(authorService.getAll());
         name.setCellValueFactory(cellData -> cellData.getValue().getNameProperty());
 
+        // allow to filter the table
         FilteredList<Author> filteredData = new FilteredList<>(authors, p -> true);
-
         filterInput.textProperty().addListener((observable, oldValue, newValue) -> {
-            filteredData.setPredicate(subject -> {
-                // If filter text is empty, display all authors.
-                if (newValue == null || newValue.isEmpty()) {
-                    return true;
-                }
-
-                String lowerCaseFilter = newValue.toLowerCase();
-
-                if (subject.getName().toLowerCase().contains(lowerCaseFilter)) {
-                    return true;
-                }
-                return false; // Does not match.
-            });
+            filteredData.setPredicate(subject -> newValue == null || newValue.isEmpty() || subject.getName().toLowerCase().contains(newValue.toLowerCase()));
         });
 
+        // show pane with info when author selected
         table.getSelectionModel().selectedItemProperty().addListener((observable, oldValue, author) -> {
             Animation.fadeInTransition(infoPane);
-
-            if(author != null) {
-                // TODO
+            if (author != null) {
+                nameInfo.setText(author.getName().toUpperCase());
+                String projects = "";
+                for(Mission mission: author.getMissions()) {
+                    if(!projects.contains(mission.getProject().getName())) {
+                        projects = projects + "- " + mission.getProject().getName() + NEW_LINE;
+                    }
+                }
+                projectsInfo.setText(projects);
             }
-
         });
 
+        // allow to sort the table
         SortedList<Author> sortedData = new SortedList<>(filteredData);
         sortedData.comparatorProperty().bind(table.comparatorProperty());
 
@@ -92,47 +95,64 @@ public class ManageAuthorController implements Initializable, AbstractController
         table.getSelectionModel().clearSelection();
     }
 
-    private void addAllAuthors() {
-        authors.addAll(authorService.getAll());
-    }
-
+    /**
+     * Show the new author screen with the info of the selected author pre-loaded
+     */
     @FXML
     private void editAuthor() {
         NewAuthorController newAuthorController = VistaNavigator.openModal(Constants.NEW_AUTHOR, language.get(Lang.NEWAUTHOR));
         newAuthorController.setAuthorForUpdate(table.getSelectionModel().getSelectedItem());
     }
 
+    /**
+     * Show the new author screen
+     */
     @FXML
     private void newAuthor() {
         VistaNavigator.openModal(Constants.NEW_AUTHOR, language.get(Lang.NEWAUTHOR));
     }
 
+    /**
+     * Delete the selected author from the database and from the table
+     */
     @FXML
     private void deleteAuthor() {
-        Alert alert = new Alert(Alert.AlertType.CONFIRMATION, "Are you sure?");
+        Alert alert = new Alert(Alert.AlertType.CONFIRMATION, language.get(Lang.CONFIRMATION));
         Optional<ButtonType> result = alert.showAndWait();
-        if (result.get() == ButtonType.OK) {
+        if (result.isPresent() && result.get() == ButtonType.OK) {
             Author author = table.getSelectionModel().getSelectedItem();
             authorService.delete(author);
             authors.remove(author);
+
+            logger.debug("Author removed: " + author);
         }
     }
 
+    /**
+     * Reload Author on edit/create
+     * @param object object to reload
+     */
     @Override
     public void reload(Object object) {
-        if(object instanceof Author) {
-            if(!authors.contains((Author) object)) {
+        if (object instanceof Author) {
+            if (!authors.contains(object)) {
                 authors.add((Author) object);
             }
-            table.getColumns().get(0).setVisible(false);
+            table.getColumns().get(0).setVisible(false);    // needed to refresh the table
             table.getColumns().get(0).setVisible(true);
+            table.getSelectionModel().clearSelection();
             table.getSelectionModel().select((Author) object);
+            logger.debug("Author reloaded: " + object);
         }
     }
 
     @Override
     public void translate() {
-
+        name.setText(language.get(Lang.NAME_COLUMN));
+        projectsLabel.setText(language.get(Lang.PARTICIPATE_IN_PROJECTS));
+        editAuthorButton.setText(language.get(Lang.EDIT));
+        newElementButton.setText(language.get(Lang.NEWAUTHOR));
+        deleteAuthorButton.setText(language.get(Lang.DELETE));
     }
 
 }
